requirements.txt
discord.py>=2.3.0
yfinance>=0.2.18
pandas>=1.5.0
numpy>=1.21.0
ta-lib>=0.4.24